Verify that users can change their passwords. Changing the password requires the current and a new password;
Verify that the passwords submitted during a registration, login, and password change are checked against a set of breached passwords. If the password is breached, the application must require users to set a new non-breached password.

Verify that passwords are stored in a form that is resistant to offline attacks. Passwords must be salted and hashed using an approved one-way key derivation or a password hashing function;